There are methods in XMLSerializer.java with names starting deserialize and do the functionalities of deserialization. These methods have been extracted to a new class with a name XMLDeserializer.
